refactor: fetchWrapper
- splits into two modules: fetch-wrapper and api-error
- renames files using kebab-case
- pull logic from imgix-api to determine loading env into fetch-wrapper
